[APIS-887] Enhancement base.py
http://jira.cubrid.org/browse/APIS-887
